+++ backend/dataall/db/models/__init__.py
@@ -5,7 +5,6 @@ from .Dashboard import Dashboard
 from .DashboardShare import DashboardShare
 from .DashboardShare import DashboardShareStatus
 from .Dataset import Dataset
-from .DatasetProfilingRun import DatasetProfilingRun
 from .DatasetQualityRule import DatasetQualityRule
 from .DatasetStorageLocation import DatasetStorageLocation
 from .DatasetTable import DatasetTable
