+++ backend/dataall/modules/datasets/db/models.py
@@ -1,7 +1,6 @@
 from sqlalchemy import Column, String
-
-from dataall.db import Base
-from dataall.db import Resource, utils
+from sqlalchemy.dialects.postgresql import JSON
+from dataall.db import Base, Resource, utils
 
 
 class DatasetTableColumn(Resource, Base):
@@ -21,3 +20,19 @@ class DatasetTableColumn(Resource, Base):
 
     def uri(self):
         return self.columnUri
+
+
+class DatasetProfilingRun(Resource, Base):
+    __tablename__ = 'dataset_profiling_run'
+    profilingRunUri = Column(
+        String, primary_key=True, default=utils.uuid('profilingrun')
+    )
+    datasetUri = Column(String, nullable=False)
+    GlueJobName = Column(String)
+    GlueJobRunId = Column(String)
+    GlueTriggerSchedule = Column(String)
+    GlueTriggerName = Column(String)
+    GlueTableName = Column(String)
+    AwsAccountId = Column(String)
+    results = Column(JSON, default={})
+    status = Column(String, default='Created')
